fix MS Windows/64bit related warnings

+++ src/headers/tomcrypt_math.h
@@ -30,7 +30,7 @@
    #define LTC_MILLER_RABIN_REPS    35
 #endif
 
-int radix_to_bin(const void *in, int radix, void *out, size_t* len);
+int radix_to_bin(const void *in, int radix, void *out, unsigned long *len);
 
 /** math descriptor */
 typedef struct {
